fix: apply dtype casting in test section to match training loop
- Add .float() cast when passing base_hidden to columns
- Add .half() cast when passing column_out to output_projection
- Fixes RuntimeError: mat1 and mat2 must have the same dtype
- Test section now matches type casting pattern from training loop
- Verified with batch size 32 on 100 samples

+++ pretraining.py
@@ -993,8 +993,8 @@ def main():
 
             # Test each column
             for col_idx in range(4):
-                column_out = columns[col_idx](base_hidden)
-                column_logits = output_projection(column_out)
+                column_out = columns[col_idx](base_hidden.float())
+                column_logits = output_projection(column_out.half())
 
                 col_pred = torch.argmax(column_logits[0, -1, :])
                 col_token = tokenizer.decode([col_pred])
